+++ CHANGES.rst
@@ -11,11 +11,17 @@ Version 0.22
 
 Unreleased
 
+* Add the :attr:`Entry.feed_url` attribute. (`#159`_)
+* Rename the :class:`EntrySearchResult` ``feed`` attribute to
+  :attr:`~EntrySearchResult.feed_url`.
+  Using ``feed`` will raise a deprecation warning in version 0.22,
+  and will be removed in the following version. (`#159`_)
 * Use ``executemany()`` instead of ``execute()`` in the SQLite storage.
   Makes updating feeds (excluding network calls) 5-10% faster. (`#144`_)
 * In the web app, redirect to the feed's page after adding a feed. (`#119`_)
 * In the web app, show highlighted search result snippets. (`#122`_)
 
+.. _#159: https://github.com/lemon24/reader/issues/159
 .. _#144: https://github.com/lemon24/reader/issues/144
 .. _#119: https://github.com/lemon24/reader/issues/119
 .. _#122: https://github.com/lemon24/reader/issues/122